+++ allensdk/__init__.py
@@ -37,7 +37,7 @@ import logging
 
 
 
-__version__ = '2.2.0'
+__version__ = '2.3.0'
 
 
 try: