+++ allensdk/brain_observatory/behavior/behavior_data_session.py
@@ -19,6 +19,8 @@ class BehaviorDataSession(object):
         self._rewards = None
         self._running_data_df = None
         self._running_speed = None
+        self._raw_running_data_df = None
+        self._raw_running_speed = None
         self._stimulus_presentations = None
         self._stimulus_templates = None
         self._stimulus_timestamps = None
@@ -93,16 +95,38 @@ class BehaviorDataSession(object):
         self._rewards = value
 
     @property
+    def raw_running_data_df(self) -> pd.DataFrame:
+        """Get running speed data without the 10Hz low pass filter.
+
+            Returns
+            -------
+            pd.DataFrame
+                Dataframe containing various signals used to compute running
+                speed, and the unfiltered speed.
+        """
+        if self._raw_running_data_df is None:
+            self._raw_running_data_df = self.api.get_running_data_df(
+                lowpass=False)
+        return self._raw_running_data_df
+
+    @raw_running_data_df.setter
+    def raw_running_data_df(self, value):
+        self._raw_running_data_df = value
+
+    @property
     def running_data_df(self) -> pd.DataFrame:
-        """Get running speed data.
+        """Get running speed data. By default applies a 10Hz low pass
+        filter to the data. To get the running speed without the filter,
+        use `raw_running_data_df`.
 
         Returns
         -------
         pd.DataFrame
-            Dataframe containing various signals used to compute running speed.
+            Dataframe containing various signals used to compute running
+            speed, and the filtered speed.
         """
         if self._running_data_df is None:
-            self._running_data_df = self.api.get_running_data_df()
+            self._running_data_df = self.api.get_running_data_df(lowpass=True)
         return self._running_data_df
 
     @running_data_df.setter
@@ -112,7 +136,8 @@ class BehaviorDataSession(object):
     @property
     def running_speed(self) -> RunningSpeed:
         """Get running speed using timestamps from
-        self.get_stimulus_timestamps.
+        self.get_stimulus_timestamps. Applies a 10Hz filter to the 
+        speed. To get the unfiltered speed, use `raw_running_speed`.
 
         NOTE: Do not correct for monitor delay.
 
@@ -125,12 +150,35 @@ class BehaviorDataSession(object):
                 Running speed of the experimental subject (in cm / s).
         """
         if self._running_speed is None:
-            self._running_speed = self.api.get_running_speed()
+            self._running_speed = self.api.get_running_speed(lowpass=True)
         return self._running_speed
 
     @running_speed.setter
     def running_speed(self, value):
         self._running_speed = value
+
+    @property
+    def raw_running_speed(self) -> RunningSpeed:
+        """Get running speed using timestamps from
+        self.get_stimulus_timestamps. Does not apply the 10Hz filter.
+
+        NOTE: Do not correct for monitor delay.
+
+        Returns
+        -------
+        RunningSpeed (NamedTuple with two fields)
+            timestamps : np.ndarray
+                Timestamps of running speed data samples
+            values : np.ndarray
+                Running speed of the experimental subject (in cm / s).
+        """
+        if self._raw_running_speed is None:
+            self._raw_running_speed = self.api.get_running_speed(lowpass=False)
+        return self._raw_running_speed
+
+    @raw_running_speed.setter
+    def raw_running_speed(self, value):
+        self._raw_running_speed = value
 
     @property
     def stimulus_presentations(self) -> pd.DataFrame: