+++ allensdk/internal/api/behavior_ophys_api.py
@@ -49,6 +49,39 @@ class BehaviorOphysLimsApi(OphysLimsApi, BehaviorOphysApiBase):
                             .corrected_stim_timestamps)
         return timestamps
 
+    @staticmethod
+    def _process_ophys_plane_timestamps(
+            ophys_timestamps: np.ndarray, plane_group: Optional[int],
+            group_count: int):
+        """
+        On mesoscope rigs each frame corresponds to a different imaging plane;
+        the laser moves between N pairs of planes. So, every Nth 2P
+        frame time in the sync file corresponds to a given plane (and
+        its multiplexed pair). The order in which the planes are
+        acquired dictates which timestamps should be assigned to which
+        plane pairs. The planes are acquired in ascending order, where
+        plane_group=0 is the first group of planes.
+
+        If the plane group is None (indicating it does not belong to
+        a plane group), then the plane was not collected concurrently
+        and the data do not need to be resampled. This is the case for
+        Scientifica 2p data, for example.
+
+        Parameters
+        ----------
+        ophys_timestamps: np.ndarray
+            Array of timestamps for 2p data
+        plane_group: int
+            The plane group this experiment belongs to. Signals the
+            order of acquisition.
+        group_count: int
+            The total number of plane groups acquired.
+        """
+        if (group_count == 0) or (plane_group is None):
+            return ophys_timestamps
+        resampled = ophys_timestamps[plane_group::group_count]
+        return resampled
+
     @memoize
     def get_ophys_timestamps(self):
 
@@ -63,6 +96,12 @@ class BehaviorOphysLimsApi(OphysLimsApi, BehaviorOphysApiBase):
         else:
             raise RuntimeError('dff_frames is longer than timestamps')
 
+        # Resample if collecting multiple concurrent planes (e.g. mesoscope)
+        plane_group = self.get_imaging_plane_group()
+        if plane_group is not None:
+            group_count = self.get_plane_group_count()
+            ophys_timestamps = self._process_ophys_plane_timestamps(
+                ophys_timestamps, plane_group, group_count)
         return ophys_timestamps
 
     @memoize
@@ -118,6 +157,7 @@ class BehaviorOphysLimsApi(OphysLimsApi, BehaviorOphysApiBase):
         metadata['LabTracks_ID'] = self.get_external_specimen_name()
         metadata['full_genotype'] = self.get_full_genotype()
         metadata['behavior_session_uuid'] = uuid.UUID(self.get_behavior_session_uuid())
+        metadata["imaging_plane_group"] = self.get_imaging_plane_group()
 
         return metadata
 
@@ -132,15 +172,15 @@ class BehaviorOphysLimsApi(OphysLimsApi, BehaviorOphysApiBase):
         return df
 
     @memoize
-    def get_running_data_df(self):
+    def get_running_data_df(self, lowpass=True):
         stimulus_timestamps = self.get_stimulus_timestamps()
         behavior_stimulus_file = self.get_behavior_stimulus_file()
         data = pd.read_pickle(behavior_stimulus_file)
-        return get_running_df(data, stimulus_timestamps)
+        return get_running_df(data, stimulus_timestamps, lowpass=lowpass)
 
     @memoize
-    def get_running_speed(self):
-        running_data_df = self.get_running_data_df()
+    def get_running_speed(self, lowpass=True):
+        running_data_df = self.get_running_data_df(lowpass=lowpass)
         assert running_data_df.index.name == 'timestamps'
         return RunningSpeed(timestamps=running_data_df.index.values,
                             values=running_data_df.speed.values)
@@ -155,7 +195,8 @@ class BehaviorOphysLimsApi(OphysLimsApi, BehaviorOphysApiBase):
         idx_name = stimulus_presentations_df_pre.index.name
         stimulus_index_df = stimulus_presentations_df_pre.reset_index().merge(stimulus_metadata_df.reset_index(), on=['image_name']).set_index(idx_name)
         stimulus_index_df.sort_index(inplace=True)
-        stimulus_index_df = stimulus_index_df[['image_set', 'image_index', 'start_time']].rename(columns={'start_time': 'timestamps'})
+        stimulus_index_df = stimulus_index_df[['image_set', 'image_index', 'start_time',
+                                               'phase', 'spatial_frequency']].rename(columns={'start_time': 'timestamps'})
         stimulus_index_df.set_index('timestamps', inplace=True, drop=True)
         stimulus_presentations_df = stimulus_presentations_df_pre.merge(stimulus_index_df, left_on='start_time', right_index=True, how='left')
         assert len(stimulus_presentations_df_pre) == len(stimulus_presentations_df)
@@ -299,7 +340,8 @@ class BehaviorOphysLimsApi(OphysLimsApi, BehaviorOphysApiBase):
         eye_tracking_data = load_eye_tracking_hdf(filepath)
         frame_times = sync_utilities.get_synchronized_frame_times(
             session_sync_file=sync_path,
-            sync_line_label_keys=Dataset.EYE_TRACKING_KEYS)
+            sync_line_label_keys=Dataset.EYE_TRACKING_KEYS,
+            trim_after_spike=False)
 
         eye_tracking_data = process_eye_tracking_data(eye_tracking_data,
                                                       frame_times,