+++ allensdk/test/brain_observatory/behavior/test_write_nwb_behavior_ophys.py
@@ -5,7 +5,7 @@ import math
 
 import allensdk.brain_observatory.nwb as nwb
 from allensdk.brain_observatory.behavior.behavior_ophys_api.behavior_ophys_nwb_api import BehaviorOphysNwbApi
-from allensdk.brain_observatory.behavior.schemas import OphysBehaviorMetaDataSchema, OphysBehaviorTaskParametersSchema
+from allensdk.brain_observatory.behavior.schemas import OphysBehaviorMetadataSchema, BehaviorTaskParametersSchema
 
 
 @pytest.mark.parametrize('roundtrip', [True, False])