+++ allensdk/test/brain_observatory/sync_utilities/test_sync_utilities.py
@@ -54,3 +54,63 @@ def test_get_synchronized_frame_times(monkeypatch, mock_dataset_fixture,
 
     obtained = su.get_synchronized_frame_times("dummy_path", sync_line_label_keys)
     assert np.allclose(obtained, expected)
+
+
+@pytest.mark.parametrize("mock_dataset_fixture,sync_line_label_keys,expected", [
+    ({"eye_tracking_timings": [0.020, 0.030, 0.040, 0.050, 3.0]},
+     Dataset.EYE_TRACKING_KEYS, [0.020, 0.030, 0.040, 0.050, 3.0]),
+
+    ({"behavior_tracking_timings": [0.080, 0.090, 0.100, 0.110, 8.0]},
+     Dataset.BEHAVIOR_TRACKING_KEYS, [0.08, 0.090, 0.100, 0.110, 8.0])
+], indirect=["mock_dataset_fixture"])
+def test_get_synchronized_frame_times_no_trim(monkeypatch, mock_dataset_fixture,
+                                      sync_line_label_keys, expected):
+    monkeypatch.setattr(su, "Dataset", mock_dataset_fixture)
+
+    obtained = su.get_synchronized_frame_times("dummy_path", sync_line_label_keys, trim_after_spike=False)
+    assert np.allclose(obtained, expected)
+
+
+@pytest.mark.parametrize("mock_dataset_fixture,sync_line_label_keys,expected", [
+    ({"eye_tracking_timings": [0.020, 0.030, 3.0, 0.040, 0.050, 0.040]},
+     Dataset.EYE_TRACKING_KEYS, [0.020, 0.030]),
+
+    ({"behavior_tracking_timings": [0.080, 8.0, 0.090, 0.100, 0.110, 0.150, 0.085, 0.110, 0.13]},
+     Dataset.BEHAVIOR_TRACKING_KEYS, [0.08])
+], indirect=["mock_dataset_fixture"])
+def test_get_synchronized_frame_times_trim_with_spike(monkeypatch, mock_dataset_fixture,
+                                      sync_line_label_keys, expected):
+    monkeypatch.setattr(su, "Dataset", mock_dataset_fixture)
+
+    obtained = su.get_synchronized_frame_times("dummy_path", sync_line_label_keys)
+    assert np.allclose(obtained, expected)
+
+
+@pytest.mark.parametrize("mock_dataset_fixture,sync_line_label_keys,expected", [
+    ({"eye_tracking_timings": [3.0, 0.030, 0.040, 0.050]},
+     Dataset.EYE_TRACKING_KEYS, []),
+
+    ({"behavior_tracking_timings": [8.0, 0.080, 0.090, 0.100, 0.110]},
+     Dataset.BEHAVIOR_TRACKING_KEYS, [])
+], indirect=["mock_dataset_fixture"])
+def test_get_synchronized_frame_times_trim_all(monkeypatch, mock_dataset_fixture,
+                                      sync_line_label_keys, expected):
+    monkeypatch.setattr(su, "Dataset", mock_dataset_fixture)
+
+    obtained = su.get_synchronized_frame_times("dummy_path", sync_line_label_keys)
+    assert np.allclose(obtained, expected)
+
+
+@pytest.mark.parametrize("mock_dataset_fixture,sync_line_label_keys,expected", [
+    ({"eye_tracking_timings": [0.020, 0.030, 3.0, 0.050, 0.040]},
+     Dataset.EYE_TRACKING_KEYS, [0.020, 0.030, 3.0, 0.050, 0.040]),
+
+    ({"behavior_tracking_timings": [0.080, 8.0, 0.090, 0.100, 0.110]},
+     Dataset.BEHAVIOR_TRACKING_KEYS, [0.080, 8.0, 0.090, 0.100, 0.110])
+], indirect=["mock_dataset_fixture"])
+def test_get_synchronized_frame_times_no_trim_with_spike(monkeypatch, mock_dataset_fixture,
+                                      sync_line_label_keys, expected):
+    monkeypatch.setattr(su, "Dataset", mock_dataset_fixture)
+
+    obtained = su.get_synchronized_frame_times("dummy_path", sync_line_label_keys, trim_after_spike=False)
+    assert np.allclose(obtained, expected)